software/scripts/uart_packet_send/requirements.txt
pyserial==3.5
bcrypt